requirements.txt

certifi==2022.12.7
cffi==1.15.1

Django==4.1.4
django-environ==0.9.0
django-social-auth==0.7.28
django-taggit==3.1.0
django_debug_toolbar==3.8.1


Pillow==9.3.0
psycopg2==2.9.5


requests==2.28.1

